The position map was 8 bytes a byte to carry one fact every 272
Two audits, one measurement and one design, and the second found that the views
are ~76% of this package's allocation while splicing is 0.2x. The map was the
cost, not the work.

# int32, then run-length

`pos`/`end` were `[]int` holding one entry per view byte. They are offsets within
one value, so they became `[]int32` — and then a run list, because measured
across the fifteen captured pages the map carries one discontinuity every 272
view bytes. A run is 12 bytes however long it is.

Three fields, not four: a run's length is the next run's start, so storing it
would be storing the same number twice. The crossover against a dense map is 1.5
view bytes per run.

    real corpus, byte-weighted   44.8x -> 26.6x -> 15.2x
    every view fires              391x ->  217x ->  209x
    with a tab                    445x ->  248x ->  230x
    css escapes beside ampersands 172x ->   96x ->   75x
    ampersands alone              118x ->   62x ->   24x
    references and brackets       104x ->   61x ->   75x

Absolute, one request on one core: a 903 KB page 36 ms and 7.6 MB held; a 1 MB
REST collection 57 ms and 13 MB; a 1.4 MB form post 19 ms and 28 MB. Concurrent:
32 page loads peak at 69 MB, eight 3.5 MB form posts at 608 MB — browsing is
free, large request bodies are the cost, because there the whole buffer is one
value and every view spans it.

**It is not a latency win.** `allocs/op` is unchanged; interleaved runs on one
core measure -5% at the minimum and +0.2% at the median. An earlier claim of a
12.6% speedup here was a lucky pair of runs, and the measurement audit refuted
it. Memory and concurrency are the win.

# The one fixture that costs more, and why it is left

`&#91;http:` repeated is the worst case for any run encoding: every reference
decodes to one view byte five source bytes wide, so no two bytes are contiguous
and every one is its own run. 61x became 75x. Real pages measure 272 bytes a
run; a body made entirely of reference-encoded brackets is a fixture, not a page.
Its ceiling is raised with that reason written down, and the other four came
down to what the code now costs.

# How the arithmetic was protected

This decides where a splice lands, and a wrong offset does not fail loudly: one
call site checks `from < 0`, the rest either panic on a slice bound or silently
skip a host, which is §4.3 with no log line.

- The accessors landed first, over the dense storage, as a verified no-op.
- The dense map is recorded *at append time* under a test-only flag, so the
  oracle shares nothing with the run arithmetic. Deriving it through the
  accessors instead would have compared the compression against itself — the
  first version did exactly that.
- 63,630 view builds: every builder x three control modes x fifteen real pages,
  the escape families beside their terminators, and 3,000 randomised inputs drawn
  from the alphabet the decoders branch on.

`TestTheViewMapAnswersWhatADenseMapWould` is the whole safety argument. If it is
ever deleted, the compression has to be reverted with it.

# Two instrument fixes from the audits

`--max-body` was unvalidated and caps the JSON and text/XML *response* paths as
well as request bodies, each of which hands the whole buffer to a view. It is
clamped to MaxInt32 now, which is what makes the int32 offsets a guarantee rather
than a comment about the 8 MiB default — and `hostshift rewrite` reads stdin with
no cap at all.

`TestAllocationStaysBounded` printed `HeapSys`, a process-monotonic high-water
mark, so three of five subtests reported the *first* one's figure. It reports the
heap actually held across the call. The one number that test printed and did not
assert on was the contaminated one.

+++ internal/rewrite/audit_r45_test.go
@@ -256,19 +256,19 @@ func stripForJSONUnicodeR45(v []byte, mode ctlMode) normalised {
 		return stripForURL(v, mode)
 	}
 	dec := make([]byte, 0, len(v))
-	pos := make([]int, 0, len(v))
-	end := make([]int, 0, len(v))
+	pos := make([]int32, 0, len(v))
+	end := make([]int32, 0, len(v))
 	for i := 0; i < len(v); {
 		if c, k := jsonUnicodeAtR45(v[i:]); k > 0 {
 			dec = append(dec, c)
-			pos = append(pos, i)
-			end = append(end, i+k)
+			pos = append(pos, int32(i))
+			end = append(end, int32(i+k))
 			i += k
 			continue
 		}
 		dec = append(dec, v[i])
-		pos = append(pos, i)
-		end = append(end, i+1)
+		pos = append(pos, int32(i))
+		end = append(end, int32(i+1))
 		i++
 	}
 	return stripRemovals(dec, pos, end, mode)
